==> notebooks/01_Web_Scraper.ipynb <==
# Initial Webscraping and Initial EDA 
---

 ***Warning: This notebook has long outputs (for the scrapers)***


import requests
import time
import pandas as pd

# Headers per reddit's requested format

headers = {'User-agent': 'Windows: class project (by /u/silentassassin82)'}

# scraping r/ask_historians

ask_hist_posts = []
after = None

for i in range(80):  
    
    if after is None:
        params={}
    else:
        params={'after': after}
    url= 'https://www.reddit.com/r/AskHistorians.json'
    res= requests.get(url, params=params, headers=headers)
    if res.status_code == 200:
        the_json= res.json()
#         posts.extend(the_json['data']['children'])
        for post in the_json['data']['children']:
            ask_hist_posts.append(post['data'])
            
        after= the_json['data']['after']
        
    else:
        print(res.status_code)
        break
        
    print('--------------------------------------')
    print('Current Scrape:')
    print(f"{len(the_json['data']['children'])} posts scraped for a total of {len(ask_hist_posts)}")
    print(f"Current after: {after}")    

    time.sleep(1)

### Converting ask_historians_posts to a DataFrame

ask_hist_df= pd.DataFrame(ask_hist_posts)
ask_hist_df['timestamp'] = pd.to_datetime(ask_hist_df['created_utc'], unit='s')

ask_hist_df.shape

# Scraping r/history posts

hist_posts = []
after = None

for i in range(80):  # _ means we don't need this variable, we just need the for loop to run 4 times.
    
    if after is None:
        params={}
    else:
        params={'after': after}
    url= 'https://www.reddit.com/r/history.json'
    res= requests.get(url, params=params, headers=headers)
    if res.status_code == 200:
        the_json= res.json()
#         posts.extend(the_json['data']['children'])
        for post in the_json['data']['children']:
            hist_posts.append(post['data'])
            
        after= the_json['data']['after']
        
    else:
        print(res.status_code)
        break
        
    print('--------------------------------------')
    print('Current Scrape:')
    print(f"{len(the_json['data']['children'])} posts scraped for a total of {len(hist_posts)}")
    print(f"Current after: {after}")    

    time.sleep(1)

hist_df= pd.DataFrame(hist_posts)
hist_df['timestamp'] = pd.to_datetime(hist_df['created_utc'], unit='s')

hist_df.shape

# Save raw data from scrapes 
hist_df.to_csv('../data/hist_scrape_2019_07_10.csv', index=False)
ask_hist_df.to_csv('../data/ask_scrape_2019_07_10.csv', index=False)

### r/History had 2 more columns than r/AskHistorians

Checking to see what those columns are

ask_columns = set(ask_hist_df.columns)  # Convert each DF into a set because each element is unique
hist_columns = set(hist_df.columns)

extra_hist= list(hist_columns - ask_columns) # Subtract out the similar columns in ask historians
print(extra_hist)

Clearly irrelevant data so they can be dropped to make sure both df's are the same dimension

hist_df.drop(columns=extra_hist, inplace=True) 

### Looking at the columns and finding most important features.

Note: Significantly more posts on r/AskHistorians, the oldest post scrape is from beginning of July.  r/History goes all the way back to May

list(ask_hist_df.columns)

Since I'm looking to classify based on text and explore how the text relates to each subreddit, I'll only pull out the posts' text and selftext and a few other columns that could be helpful or interesting.  They may not be necessary but I'd prefer to carry them along than have to go back and get them.

good_cols = ['subreddit', 'author', 'locked', 'num_comments', 'selftext', 'title', 'timestamp'] 

AskHistorians = ask_hist_df[good_cols] 
History = hist_df[good_cols]

If I want to used whether a post was locked or not later on, I'll need them to be in numerical format so I'll do a simple one-hot encoding.

History['locked']= History['locked'].map({False: 0, 
                       True: 1})

AskHistorians['locked']= AskHistorians['locked'].map({False: 0, 
                       True: 1})

Save the raw dataframes to a csv for use in the next notebook (and avoid the copy of a dataframe error)

AskHistorians.to_csv("../data/ask_historians.csv", index=False)
History.to_csv('../data/history.csv', index=False)
