==> notebooks/02_Pre-processing.ipynb <==
#  Pre-processing
---

import pandas as pd
import numpy as np

ask_df = pd.read_csv('../data/ask_historians.csv')
hist_df = pd.read_csv('../data/history.csv')

### Cleaning and Removing Nulls

reddit_df= pd.concat([ask_df, hist_df])

reddit_df.isnull().mean()

NaNs are justs with no selftext so no reason to drop and better to fill with empty string

reddit_df.fillna('', inplace=True)

# Creating a column that includes the full text of the post; both the body and title for a self.text posts.

reddit_df['full_text']= reddit_df.eval('title + selftext')

reddit_df.isnull().sum()

reddit_df.dtypes

# Adapted from class code, cleans and lemmatizes the text.

from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords
import regex as re
def lem_str(text):
    
    lemmatizer = WordNetLemmatizer()
    
    # Remove non-letters and trailing \n + some number.
    text = re.sub(r'[^a-zA-Z0-9]', ' ', text)
    text = text.replace(r'[\\\n0-9]', ' ')
    
    # Convert to lower case, split/tokenize into individual words.
    words = text.lower().split()
    
    stops = set(stopwords.words('english'))
    
    # Remove stop words.
    meaningful_words = [w for w in words if not w in stops]
    
    lem = [lemmatizer.lemmatize(i) for i in meaningful_words]
    
    return" ".join(lem)

reddit_df['full_text_clean']= reddit_df['full_text'].map(lem_str)

reddit_df.shape

lengths_of_posts= np.array([len(post.split()) for post in reddit_df['full_text_clean']])
lengths_of_posts.mean()

**Average length of posts across all subreddits:** $47.7$ words per post

Defining my postive and negative class.  AskHistorians is positive and History is negative.

reddit_df['subreddit']= reddit_df['subreddit'].map({'AskHistorians': 1,
                           'history': 0})

reddit_df.head()

reddit_df['subreddit'].value_counts()

reddit_df.to_csv('../data/reddit_df.csv', index=False)
