==> notebooks/04_Tf-idf.ipynb <==
#  Term Frequency - Inverse Document Frequency Vectorization
---

Comparing CountVectorizer with Tf-idf

import pandas as pd

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix
from sklearn.metrics import f1_score

reddit_df= pd.read_csv('../data/reddit_df.csv')

reddit_df.head()

### Model with only text and tfidf

X= reddit_df['full_text_clean']
y= reddit_df['subreddit']

X_train, X_test, y_train, y_test= train_test_split(X, y, stratify=y, random_state=22)

tfi = TfidfVectorizer(ngram_range=(1,2), # Keep n_gram range consistent with cvec
                     max_df=.98,
                     min_df=2)

X_train_tfi = tfi.fit_transform(X_train) 
X_test_tfi = tfi.transform(X_test)

X_train_df = pd.DataFrame(X_train_tfi.toarray(), columns = tfi.get_feature_names())
X_test_df = pd.DataFrame(X_test_tfi.toarray(), columns = tfi.get_feature_names())

lr = LogisticRegression()

lr.fit(X_train_df, y_train)
lr.score(X_train_df, y_train)

lr.score(X_test_df, y_test)

**Train Accuracy** = $0.975$

**Test Accuracy** = $0.867$

y_pred = lr.predict(X_test_tfi)

tn, fp, fn, tp = confusion_matrix(y_test,y_pred).ravel()

**Specificity and Sensitivity**

print(f'Specificity: {round(tn / (fp + tn), 2)}')

print(f'Sensitivity: {round(tp /(tp + fn), 2)}')

**F1 Score**

round(f1_score(y_test, y_pred),3)

**Cross Val Score**

cross_val_score(lr, X_train_tfi, y_train, cv=20, n_jobs=4).mean()

### Observations

The tf-idf did significantly worse than the Countvectorizer.  My assumption would be that frequency in each document has a neglible effect because each post has a very specific event and time period and it would be the topic of the post as a whole (i.e. American Revolution/Civil War or Ancient Rome) rather than a weighted average of individual terms.
